requirements.txt
tqdm>=4.19.4
matplotlib
numpy
torchvision==0.7.0
requests==2.24.0
Flask==1.1.2
opencv_python==3.4.1.15
numpy==1.19.1
matplotlib==3.3.2
torch==1.6.0
ignite==1.1.0
Pillow==8.0.1
